decompose workflow of kratos into 4 substeps
